+++ packages/extension/lib/extension.dart
@@ -1,4 +1,4 @@
-export 'date.dart';
-export 'enum.dart';
-export 'string.dart';
-export 'list.dart';
+export 'src/date.dart';
+export 'src/enum.dart';
+export 'src/string.dart';
+export 'src/iterable.dart';