+++ packages/extension/lib/src/date.dart
@@ -1,4 +1,35 @@
 extension DateUtils on DateTime {
+  /// Tomorrow at same hour / minute / second than now
+  static DateTime get tomorrow => DateTime.now().nextDay;
+
+  /// Yesterday at same hour / minute / second than now
+  static DateTime get yesterday => DateTime.now().previousDay;
+
+  /// Current date (Same as [Date.now])
+  static DateTime get today => DateTime.now();
+
+  /// Returns a [DateTime] for each day the given range.
+  ///
+  /// [start] inclusive
+  /// [end] exclusive
+  static Iterable<DateTime> daysInRange(DateTime start, DateTime end) sync* {
+    var i = start;
+    var offset = start.timeZoneOffset;
+    while (i.isBefore(end)) {
+      yield i;
+      i = i.addDays(1);
+      var timeZoneDiff = i.timeZoneOffset - offset;
+      if (timeZoneDiff.inSeconds != 0) {
+        offset = i.timeZoneOffset;
+        i = i.subtract(Duration(seconds: timeZoneDiff.inSeconds));
+      }
+    }
+  }
+
+  /// Returns a [DateTime] with the date of the original, but time set to
+  /// midnight.
+  DateTime get dateOnly => DateTime(this.year, this.month, this.day);
+
   bool get isToday {
     final nowDate = DateTime.now();
     return year == nowDate.year && month == nowDate.month && day == nowDate.day;
@@ -18,6 +49,40 @@ extension DateUtils on DateTime {
         day == nowDate.day + 1;
   }
 
+  /// Add a certain amount of days to this date
+  DateTime addDays(int amount) => DateTime(
+        year,
+        month,
+        day + amount,
+        hour,
+        minute,
+        second,
+        millisecond,
+        microsecond,
+      );
+
+  /// Add a certain amount of hours to this date
+  DateTime addHours(int amount) => DateTime(
+        year,
+        month,
+        day,
+        hour + amount,
+        minute,
+        second,
+        millisecond,
+        microsecond,
+      );
+
+  /// The day after this [DateTime]
+  DateTime get nextDay => addDays(1);
+
+  /// The day previous this [DateTime]
+  DateTime get previousDay => addDays(-1);
+
+  /// Whether or not two times are on the same day.
+  bool isSameDay(DateTime b) =>
+      year == b.year && month == b.month && day == b.day;
+
   /// The list of days in a given month
   List<DateTime> get daysInMonth {
     var first = firstDayOfMonth;
@@ -36,9 +101,9 @@ extension DateUtils on DateTime {
     return daysInRange(firstToDisplay, lastToDisplay).toList();
   }
 
-  bool get isFirstDayOfMonth => isSameDay(firstDayOfMonth, this);
+  bool get isFirstDayOfMonth => isSameDay(firstDayOfMonth);
 
-  bool get isLastDayOfMonth => isSameDay(lastDayOfMonth, this);
+  bool get isLastDayOfMonth => isSameDay(lastDayOfMonth);
 
   DateTime get firstDayOfMonth => DateTime(this.year, this.month);
 
@@ -101,42 +166,20 @@ extension DateUtils on DateTime {
 
   DateTime get nextWeek => this.add(Duration(days: 7));
 
-  /// Returns a [DateTime] for each day the given range.
-  ///
-  /// [start] inclusive
-  /// [end] exclusive
-  static Iterable<DateTime> daysInRange(DateTime start, DateTime end) sync* {
-    var i = start;
-    var offset = start.timeZoneOffset;
-    while (i.isBefore(end)) {
-      yield i;
-      i = i.add(Duration(days: 1));
-      var timeZoneDiff = i.timeZoneOffset - offset;
-      if (timeZoneDiff.inSeconds != 0) {
-        offset = i.timeZoneOffset;
-        i = i.subtract(Duration(seconds: timeZoneDiff.inSeconds));
-      }
-    }
-  }
-
-  static bool isSameWeek(DateTime a, DateTime b) {
+  bool isSameWeek(DateTime b) {
     /// Handle Daylight Savings by setting hour to 12:00 Noon
     /// rather than the default of Midnight
-    a = DateTime.utc(a.year, a.month, a.day);
+    final a = DateTime.utc(year, month, day);
     b = DateTime.utc(b.year, b.month, b.day);
 
-    var diff = a.toUtc().difference(b.toUtc()).inDays;
+    final diff = a.toUtc().difference(b.toUtc()).inDays;
     if (diff.abs() >= 7) {
       return false;
     }
 
-    var min = a.isBefore(b) ? a : b;
-    var max = a.isBefore(b) ? b : a;
-    var result = max.weekday % 7 - min.weekday % 7 >= 0;
+    final min = a.isBefore(b) ? a : b;
+    final max = a.isBefore(b) ? b : a;
+    final result = max.weekday % 7 - min.weekday % 7 >= 0;
     return result;
   }
-
-  /// Whether or not two times are on the same day.
-  static bool isSameDay(DateTime a, DateTime b) =>
-      a.year == b.year && a.month == b.month && a.day == b.day;
 }