+++ packages/extension/test/all.dart
@@ -1,7 +1,9 @@
 import './enum.dart' as _enum;
 import './string.dart' as _string;
+import './date.dart' as _date;
 
 main() {
   _enum.main();
   _string.main();
+  _date.main();
 }